requirements.txt
Django==5.2.11
Pygments==2.19.2
RTFDE==0.1.2.2
beautifulsoup4==4.13.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
crispy-bootstrap5==2025.6
cryptography==46.0.4
django-allauth==65.14.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-debug-toolbar==6.2.0
django-tinymce==5.0.0
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.55.0
idna==3.11
iniconfig==2.3.0
lark==1.3.1
msoffcrypto-tool==6.0.0
mssql-django==1.6
olefile==0.47
oletools==0.60.2
packaging==26.0
pcodedmp==1.2.6
pluggy==1.6.0
pycparser==3.0
pyodbc==5.3.0
pyparsing==3.3.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
red-black-tree-mod==1.22
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
win_unicode_console==0.5
asgiref==3.11.1
